Update topic info after create/update ops

+++ kafka-ui-api/src/main/java/com/provectus/kafka/ui/cluster/model/InternalClusterMetrics.java
@@ -22,7 +22,7 @@ public class InternalClusterMetrics {
     private final int outOfSyncReplicasCount;
     private final Map<String, Number> bytesInPerSec;
     private final Map<String, Number> bytesOutPerSec;
-    private final int segmentCount;
+    private final long segmentCount;
     private final long segmentSize;
     private final Map<Integer, InternalBrokerMetrics> internalBrokerMetrics;
     private final List<Metric> metrics;
